refactor(string-calculator): create Splitter Interface
