+++ src/main/java/calculator/splitter/Delimiter.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.splitter;
 
 public class Delimiter {
