+++ src/main/java/calculator/splitter/Splitter.java
@@ -1,0 +1,7 @@
+package calculator.splitter;
+
+@FunctionalInterface
+public interface Splitter<V> {
+
+    V[] split(V target);
+}
